Refactored plot + processing funcs into libraries
